F1000Research/scripts/requirements.txt
python-dotenv
ipykernel>=6.13.0
pytest>=7.1.2
matplotlib>=3.5.1
scipy>=1.8.0
scikit-learn>=1.0.2
pandas>=1.4.2
matplotlib>=3.5.1